+++ console/src/features/FeatureFlags/availableFeatureFlags.ts
@@ -1,9 +1,11 @@
 import { FeatureFlagDefinition } from './types';
 
 const relationshipTabTablesId = '0bea35ff-d3e9-45e9-af1b-59923bf82fa9';
+const addRemoteSchemaId = 'bf57c2ba-cab2-11ec-9d64-0242ac120002';
 
 export const availableFeatureFlagIds = {
   relationshipTabTablesId,
+  addRemoteSchemaId,
 };
 
 export const availableFeatureFlags: FeatureFlagDefinition[] = [
@@ -18,4 +20,14 @@ export const availableFeatureFlags: FeatureFlagDefinition[] = [
   //   defaultValue: false,
   //   discussionUrl: '',
   // },
+  {
+    id: addRemoteSchemaId,
+    title: 'New create remote schema page',
+    description:
+      'Try out the new Add Remote Schema page that supports GraphQL customization',
+    section: 'remote schemas',
+    status: 'alpha',
+    defaultValue: false,
+    discussionUrl: '',
+  },
 ];