+++ console/src/features/RemoteSchema/components/Headers/index.ts
@@ -1,0 +1,2 @@
+export * from './Headers';
+export { headersSchema } from './schema';